+++ oblate/fields/__init__.py
@@ -8,3 +8,4 @@ from oblate.fields.base import *
 from oblate.fields.primitive import *
 from oblate.fields.nesting import *
 from oblate.fields.typings import *
+from oblate.fields.structs import *